+++ src/jgo/cli/commands/lock.py
@@ -30,7 +30,7 @@ def lock(ctx, check):
       jgo lock --check
       jgo lock --update
     """
-    from ...config.jgorc import JgoConfig
+    from ...config.file import JgoConfig
     from ..parser import _build_parsed_args
 
     opts = ctx.obj