+++ src/jgo/cli/commands/remove.py
@@ -31,7 +31,7 @@ def remove(ctx, coordinates, no_sync):
       jgo remove org.scijava:scijava-common org.scijava:parsington
       jgo remove --no-sync net.imagej:imagej
     """
-    from ...config.jgorc import JgoConfig
+    from ...config.file import JgoConfig
     from ..parser import _build_parsed_args
 
     opts = ctx.obj